+++ src/main/java/org/markdownwriterfx/options/GeneralOptionsPane.java
@@ -30,11 +30,19 @@ package org.markdownwriterfx.options;
 import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
 import java.util.SortedMap;
 import javafx.scene.control.CheckBox;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.Label;
+import javafx.scene.control.ListCell;
+import javafx.scene.control.Spinner;
+import javafx.scene.control.SpinnerValueFactory;
 import javafx.scene.control.TextField;
+import javafx.scene.text.Font;
+import javafx.scene.text.FontPosture;
+import javafx.scene.text.FontWeight;
+import javafx.scene.text.Text;
 import org.markdownwriterfx.Messages;
 import org.markdownwriterfx.util.Item;
 import org.markdownwriterfx.util.Utils;
@@ -52,6 +60,20 @@ public class GeneralOptionsPane
 	public GeneralOptionsPane() {
 		initComponents();
 
+		Font titleFont = Font.font(16);
+		editorSettingsLabel.setFont(titleFont);
+		fileSettingsLabel.setFont(titleFont);
+
+		// font family
+		fontFamilyField.getItems().addAll(getMonospacedFonts());
+		fontFamilyField.getSelectionModel().select(0);
+		fontFamilyField.setButtonCell(new FontListCell());
+		fontFamilyField.setCellFactory(p -> new FontListCell());
+
+		// font size
+		fontSizeField.setValueFactory(new SpinnerValueFactory.IntegerSpinnerValueFactory(8, 36));
+
+		// line separator
 		String defaultLineSeparator = System.getProperty( "line.separator", "\n" );
 		String defaultLineSeparatorStr = defaultLineSeparator.replace("\r", "CR").replace("\n", "LF");
 		lineSeparatorField.getItems().addAll(
@@ -59,9 +81,39 @@ public class GeneralOptionsPane
 			new Item<String>( Messages.get("GeneralOptionsPane.sepWindows"), "\r\n" ),
 			new Item<String>( Messages.get("GeneralOptionsPane.sepUnix"), "\n" ));
 
+		// encoding
 		encodingField.getItems().addAll(getAvailableEncodings());
 
+		// file extensions
 		markdownFileExtensionsField.setPromptText(Options.DEF_MARKDOWN_FILE_EXTENSIONS);
+	}
+
+	/**
+	 * Return a list of all the mono-spaced fonts on the system.
+	 *
+	 * @author David D. Clark http://clarkonium.net/2015/07/finding-mono-spaced-fonts-in-javafx/
+	 */
+	private static Collection<String> getMonospacedFonts() {
+
+		// Compare the layout widths of two strings. One string is composed
+		// of "thin" characters, the other of "wide" characters. In mono-spaced
+		// fonts the widths should be the same.
+
+		final Text thinTxt = new Text("1 l"); // note the space
+		final Text thikTxt = new Text("MWX");
+
+		List<String> fontFamilyList = Font.getFamilies();
+		List<String> monospacedFonts = new ArrayList<>();
+
+		for (String fontFamilyName : fontFamilyList) {
+			Font font = Font.font(fontFamilyName, FontWeight.NORMAL, FontPosture.REGULAR, 14.0d);
+			thinTxt.setFont(font);
+			thikTxt.setFont(font);
+			if (thinTxt.getLayoutBounds().getWidth() == thikTxt.getLayoutBounds().getWidth())
+				monospacedFonts.add(fontFamilyName);
+		}
+
+		return monospacedFonts;
 	}
 
 	private Collection<Item<String>> getAvailableEncodings() {
@@ -75,25 +127,40 @@ public class GeneralOptionsPane
 	}
 
 	void load() {
+		// editor settings
+		fontFamilyField.getSelectionModel().select(Options.getFontFamily());
+		fontSizeField.getValueFactory().setValue(Options.getFontSize());
+		showWhitespaceCheckBox.setSelected(Options.isShowWhitespace());
+
+		// file settings
 		lineSeparatorField.setValue(new Item<String>(Options.getLineSeparator(), Options.getLineSeparator()));
 		encodingField.setValue(new Item<String>(Options.getEncoding(), Options.getEncoding()));
 		markdownFileExtensionsField.setText(Options.getMarkdownFileExtensions());
-
-		showWhitespaceCheckBox.setSelected(Options.isShowWhitespace());
 	}
 
 	void save() {
+		// editor settings
+		Options.setFontFamily(fontFamilyField.getSelectionModel().getSelectedItem());
+		Options.setFontSize(fontSizeField.getValue());
+		Options.setShowWhitespace(showWhitespaceCheckBox.isSelected());
+
+		// file settings
 		Options.setLineSeparator(lineSeparatorField.getValue().value);
 		Options.setEncoding(encodingField.getValue().value);
 		Options.setMarkdownFileExtensions(Utils.defaultIfEmpty(
 				markdownFileExtensionsField.getText().trim(),
 				Options.DEF_MARKDOWN_FILE_EXTENSIONS));
-
-		Options.setShowWhitespace(showWhitespaceCheckBox.isSelected());
 	}
 
 	private void initComponents() {
 		// JFormDesigner - Component initialization - DO NOT MODIFY  //GEN-BEGIN:initComponents
+		editorSettingsLabel = new Label();
+		fontFamilyLabel = new Label();
+		fontFamilyField = new ComboBox<>();
+		fontSizeLabel = new Label();
+		fontSizeField = new Spinner<>();
+		showWhitespaceCheckBox = new CheckBox();
+		fileSettingsLabel = new Label();
 		Label lineSeparatorLabel = new Label();
 		lineSeparatorField = new ComboBox<>();
 		Label lineSeparatorLabel2 = new Label();
@@ -101,52 +168,99 @@ public class GeneralOptionsPane
 		encodingField = new ComboBox<>();
 		Label markdownFileExtensionsLabel = new Label();
 		markdownFileExtensionsField = new TextField();
-		showWhitespaceCheckBox = new CheckBox();
 
 		//======== this ========
-		setCols("[fill][fill][grow,fill]");
-		setRows("[][][]para[]");
+		setCols("[indent,fill]0[fill][fill][grow,fill]");
+		setRows("[][][][]para[][][][]para");
+
+		//---- editorSettingsLabel ----
+		editorSettingsLabel.setText(Messages.get("GeneralOptionsPane.editorSettingsLabel.text"));
+		add(editorSettingsLabel, "cell 0 0 2 1");
+
+		//---- fontFamilyLabel ----
+		fontFamilyLabel.setText(Messages.get("GeneralOptionsPane.fontFamilyLabel.text"));
+		fontFamilyLabel.setMnemonicParsing(true);
+		add(fontFamilyLabel, "cell 1 1");
+
+		//---- fontFamilyField ----
+		fontFamilyField.setVisibleRowCount(20);
+		add(fontFamilyField, "cell 2 1");
+
+		//---- fontSizeLabel ----
+		fontSizeLabel.setText(Messages.get("GeneralOptionsPane.fontSizeLabel.text"));
+		fontSizeLabel.setMnemonicParsing(true);
+		add(fontSizeLabel, "cell 1 2");
+		add(fontSizeField, "cell 2 2,alignx left,growx 0");
+
+		//---- showWhitespaceCheckBox ----
+		showWhitespaceCheckBox.setText(Messages.get("GeneralOptionsPane.showWhitespaceCheckBox.text"));
+		add(showWhitespaceCheckBox, "cell 1 3 3 1,alignx left,growx 0");
+
+		//---- fileSettingsLabel ----
+		fileSettingsLabel.setText(Messages.get("GeneralOptionsPane.fileSettingsLabel.text"));
+		add(fileSettingsLabel, "cell 0 4 2 1");
 
 		//---- lineSeparatorLabel ----
 		lineSeparatorLabel.setText(Messages.get("GeneralOptionsPane.lineSeparatorLabel.text"));
 		lineSeparatorLabel.setMnemonicParsing(true);
-		add(lineSeparatorLabel, "cell 0 0");
-		add(lineSeparatorField, "cell 1 0");
+		add(lineSeparatorLabel, "cell 1 5");
+		add(lineSeparatorField, "cell 2 5");
 
 		//---- lineSeparatorLabel2 ----
 		lineSeparatorLabel2.setText(Messages.get("GeneralOptionsPane.lineSeparatorLabel2.text"));
-		add(lineSeparatorLabel2, "cell 2 0");
+		add(lineSeparatorLabel2, "cell 3 5");
 
 		//---- encodingLabel ----
 		encodingLabel.setText(Messages.get("GeneralOptionsPane.encodingLabel.text"));
 		encodingLabel.setMnemonicParsing(true);
-		add(encodingLabel, "cell 0 1");
+		add(encodingLabel, "cell 1 6");
 
 		//---- encodingField ----
 		encodingField.setVisibleRowCount(20);
-		add(encodingField, "cell 1 1");
+		add(encodingField, "cell 2 6");
 
 		//---- markdownFileExtensionsLabel ----
 		markdownFileExtensionsLabel.setText(Messages.get("GeneralOptionsPane.markdownFileExtensionsLabel.text"));
 		markdownFileExtensionsLabel.setMnemonicParsing(true);
-		add(markdownFileExtensionsLabel, "cell 0 2");
-		add(markdownFileExtensionsField, "cell 1 2 2 1");
-
-		//---- showWhitespaceCheckBox ----
-		showWhitespaceCheckBox.setText(Messages.get("GeneralOptionsPane.showWhitespaceCheckBox.text"));
-		add(showWhitespaceCheckBox, "cell 0 3 3 1,alignx left,growx 0");
+		add(markdownFileExtensionsLabel, "cell 1 7");
+		add(markdownFileExtensionsField, "cell 2 7 2 1");
 		// JFormDesigner - End of component initialization  //GEN-END:initComponents
 
 		// TODO set this in JFormDesigner as soon as it supports labelFor
+		fontFamilyLabel.setLabelFor(fontFamilyField);
+		fontSizeLabel.setLabelFor(fontSizeField);
 		lineSeparatorLabel.setLabelFor(lineSeparatorField);
 		encodingLabel.setLabelFor(encodingField);
 		markdownFileExtensionsLabel.setLabelFor(markdownFileExtensionsField);
 	}
 
 	// JFormDesigner - Variables declaration - DO NOT MODIFY  //GEN-BEGIN:variables
+	private Label editorSettingsLabel;
+	private Label fontFamilyLabel;
+	private ComboBox<String> fontFamilyField;
+	private Label fontSizeLabel;
+	private Spinner<Integer> fontSizeField;
+	private CheckBox showWhitespaceCheckBox;
+	private Label fileSettingsLabel;
 	private ComboBox<Item<String>> lineSeparatorField;
 	private ComboBox<Item<String>> encodingField;
 	private TextField markdownFileExtensionsField;
-	private CheckBox showWhitespaceCheckBox;
 	// JFormDesigner - End of variables declaration  //GEN-END:variables
+
+	//---- class FontListCell -------------------------------------------------
+
+	private static class FontListCell
+		extends ListCell<String>
+	{
+		@Override
+		protected void updateItem(String fontFamily, boolean empty) {
+			super.updateItem(fontFamily, empty);
+
+			if (!empty) {
+				setText(fontFamily);
+				setFont(Font.font(fontFamily));
+			} else
+				setText(null);
+		}
+	}
 }